+++ src/components/SendInvite/SelectPassengers.tsx
@@ -12,8 +12,8 @@ export function SelectPassengers( { passengers, setPassengers } : { passengers :
         <div className='custom-component'>
  
             <Typography variant='h6'>Acompañantes:</Typography>
-            <Card elevation={6} id='card'>
-                <TextField variant='filled' type='number' label='Ej: 3 acompañantes' className='text-fields' value={ passengers>0 ? String(passengers) : '' } autoFocus={ passengers>0 } onChange={ handlePassangers }/>
+            <Card elevation={6} id='card card-sedn-invite'>
+                <TextField variant='filled' type='number' label='Ej: 3 acompañantes' className='text-fields max-width' value={ passengers>0 ? String(passengers) : '' } autoFocus={ passengers>0 } onChange={ handlePassangers }/>
             </Card>
  
         </div>